+++ middleman-core/lib/middleman-core/renderers/coffee_script.rb
@@ -16,7 +16,6 @@ module Middleman
           ::Tilt.prefer(DebuggingCoffeeScriptTemplate)
 
           app.before_configuration do
-            template_extensions :coffee => :js
             DebuggingCoffeeScriptTemplate.middleman_app = self
           end
         end