+++ middleman-core/lib/middleman-core/renderers/markdown.rb
@@ -13,14 +13,6 @@ module Middleman
           app.config.define_setting :markdown_engine, :kramdown, 'Preferred markdown engine'
           app.config.define_setting :markdown_engine_prefix, ::Tilt, 'The parent module for markdown template engines'
 
-          app.before_configuration do
-            template_extensions :markdown => :html,
-                                :mdown    => :html,
-                                :md       => :html,
-                                :mkd      => :html,
-                                :mkdn     => :html
-          end
-
           # Once configuration is parsed
           app.after_configuration do
             markdown_exts = %w(markdown mdown md mkd mkdn)