+++ middleman-core/lib/middleman-core/renderers/stylus.rb
@@ -14,11 +14,6 @@ module Middleman
         def registered(app)
           # Default stylus options
           app.config.define_setting :styl, {}, 'Stylus config options'
-
-
-          app.before_configuration do
-            template_extensions :styl => :css
-          end
         end
 
         alias :included :registered